% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adaptive_Variable_Selection.R
\name{adaptive_variable_selection}
\alias{adaptive_variable_selection}
\title{Adaptive Variable Selection}
\usage{
adaptive_variable_selection(
  pre_trained_prs,
  validation_prs,
  training_phenotype,
  validation_phenotype,
  training_theta_encoding,
  validation_theta_encoding,
  family,
  biglasso = FALSE,
  lam.max = 0.002,
  lam.min = 6e-05,
  nlambda = 50,
  alpha = 0.1,
  tolerance = 0.025,
  threshold = 0.01,
  err = 1e-05,
  delta = NULL
)
}
\arguments{
\item{pre_trained_prs}{a dataframe containing the pre-trained training polygenic risk scores}

\item{validation_prs}{a dataframe containing the pre-trained validation polygenic risk scores}

\item{training_phenotype}{a dataframe containing the training phenotype data}

\item{validation_phenotype}{a dataframe containing the validation phenotype data}

\item{training_theta_encoding}{a dataframe containing the training thetas}

\item{validation_theta_encoding}{a dataframe containing the validation thetas}

\item{family}{a string specifying the type of model to fit a \link[stats]{glm}.
Use "binomial" for binary outcomes and "gaussian" for continuous outcomes.}

\item{biglasso}{a logical flag indicating whether to use big data optimization.
If true, then the program uses \link[biglasso]{biglasso}. Or else, it uses
the standard \link[glmnet]{glmnet} (default: FALSE)}

\item{lam.max}{maximum value of the regularization parameter (lambda) to be
considered. This value is used to generate the lambda sequence in \link[glmnet]{glmnet}.}

\item{lam.min}{minimum value of the regularization parameter (lambda) to be
considered. This value is used to generate the lambda sequence in \link[glmnet]{glmnet}.}

\item{nlambda}{number of different lambda values to be evaluated between
lam.max and lam.min. This value is used to generate the lambda sequence in \link[glmnet]{glmnet}.}

\item{alpha}{desired FDR control level.}

\item{tolerance}{tolerance level for noise in mirror statistics.}

\item{threshold}{threshold value for determining significance in mirror
statistics.}

\item{err}{error threshold for model performance improvement.}

\item{delta}{additional tuning parameter for the FDR control.}
}
\value{
A vector of selected \eqn{\theta_{SNPs}}.
}
\description{
This function identifies important \eqn{\theta_{SNPs}} through adaptive variable
selection and FDR control.
}
