% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABPRS.R
\name{apply_weights}
\alias{apply_weights}
\title{Applying Weights to Produce AB-PRSs}
\usage{
apply_weights(pre_trained_prs, genotype, weights)
}
\arguments{
\item{pre_trained_prs}{a dataframe containing the pre-trained polygenic risk scores}

\item{genotype}{a dataframe containing the corresponding genotype information,
whose rows represent individuals and columns represent SNPs.}

\item{weights}{a dataframe containing the weights of the pre-trained prs and
important \eqn{\theta_{SNPs}}, and the theta1 and theta2 values for those SNPs.
This dataframe can be generated from \link[ABPRS]{ABPRS}. Note that this function
assumes that the first row contains the pre-trained PRS while the other rows contain
information about the \eqn{\theta_{SNPs}}.}
}
\value{
A dataframe with the adaptively-boosted polygenic risk scores.
}
\description{
This function applies the weights generated from the ABPRS function to produce
adaptively-boosted polygenic risk scores.
}
