% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{AUC_test}
\alias{AUC_test}
\title{Area Under the Curve Score for Binary Outcomes}
\usage{
AUC_test(
  support,
  prs_train = prs_train,
  prs_test = prs_test,
  dat_train = dat_train,
  dat_test = dat_test,
  theta_train = theta_train,
  theta_test = theta_test,
  lasso = FALSE,
  lambda = NULL,
  delta = 0
)
}
\arguments{
\item{support}{vector of indices specifying which columns of thetas to
include in the model}

\item{prs_train}{dataframe containing the training polygenic risk scores}

\item{prs_test}{dataframe containing the testing polygenic risk scores}

\item{dat_train}{dataframe containing the training phenotype data}

\item{dat_test}{dataframe containing the testing phenotype data}

\item{theta_train}{dataframe containing the training thetas}

\item{theta_test}{dataframe containing the testing thetas}

\item{lasso}{boolean flag to determine whether to use lasso regression}

\item{lambda}{tuning parameters for lasso regression}

\item{delta}{additional tuning parameter for the Lasso regression; used for
refined debiased lasso regression (default: 0)}
}
\value{
the AUC score of the data's prediction performance
}
\description{
This function computes the Area Under the Curve (AUC) scores for the
test data.
}
