% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betas_thetas.R
\name{convert_theta}
\alias{convert_theta}
\title{Convert Theta SNPs}
\usage{
convert_theta(geno, betas)
}
\arguments{
\item{geno}{a dataframe containing genotype data with original SNP encoding}

\item{betas}{a dataframe containing beta1 and beta2 coefficients (See
\link{learning_betas} to generate the beta dataframe)}
}
\value{
A dataframe with the converted theta SNPs
}
\description{
This function converts the genotype matrix and the weighted beta coefficients
into theta SNPs by replacing homozygous common allele AA with 0, heterozygous
allele Aa with beta1, and homozygous minor allele with beta2.
}
