% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thetas.R
\name{encoding_theta_snps}
\alias{encoding_theta_snps}
\title{Convert SNPs to \eqn{\theta_{SNP}} Encoding}
\usage{
encoding_theta_snps(genotype, thetas)
}
\arguments{
\item{genotype}{a dataframe containing genotype data with additive SNP encoding.}

\item{thetas}{a dataframe containing theta encoding for genotype Aa (theta1)
and aa (theta2) (See \link{learning_theta_snps} for the function to generate
this dataframe).}
}
\value{
A dataframe containing genotype matrix with theta encoding.
}
\description{
This function converts a genotype matrix with additive encoding into a
theta-encoded matrix. It replaces the encoding of heterozygous alleles (Aa)
with theta1 and homozygous alternative alleles (aa) with theta2, while retaining
the encoding of homozygous reference alleles (AA) as 0.
}
