% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDR_ABPRS.R
\name{FDR_selection_GLM}
\alias{FDR_selection_GLM}
\title{FDR control via mirror statistics from GLM (binary phenotype)}
\usage{
FDR_selection_GLM(
  prs_train = prs_train,
  prs_val = prs_val,
  pheno_train = pheno_train,
  pheno_val = pheno_val,
  theta_train = theta_train,
  theta_val = theta_val,
  bigdata = FALSE,
  lam.max = 0.002,
  lam.min = 6e-05,
  nlambda = 50,
  alpha = 0.1,
  tolerance = 0.025,
  threshold = 0.01,
  err = 1e-06,
  delta = NULL
)
}
\arguments{
\item{prs_train}{dataframe containing the training polygenic risk scores}

\item{prs_val}{dataframe containing the validation polygenic risk scores}

\item{pheno_train}{dataframe containing the training phenotype data}

\item{pheno_val}{dataframe containing the validation phenotype data}

\item{theta_train}{dataframe containing the training thetas}

\item{theta_val}{dataframe containing the validation thetas}

\item{bigdata}{logical flag indicating whether to use big data optimization
(default: FALSE)}

\item{lam.max}{maximum value of the regularization parameter (lambda) to be
considered}

\item{lam.min}{minimum value of the regularization parameter (lambda) to be
considered}

\item{nlambda}{number of different lambda values to be evaluated between
lam.max and lam.min}

\item{alpha}{desired FDR control level}

\item{tolerance}{tolerance level for noise in mirror statistics}

\item{threshold}{threshold value for determining significance in mirror
statistics}

\item{err}{error threshold for model performance improvement}

\item{delta}{additional tuning parameter for the FDR control}
}
\value{
A large list with the following components:
\itemize{
\item support.F: selected false support
\item support.T: selected true support based on positive performance metrics for
both validation and training scores (removed later)
\item support.list: list of selected support for each lambda value.
\item training_validation_t: matrix of training and validation performance metrics (AUC score)
}
}
\description{
This function conducts FDR control through mirror statistics.
}
