% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{geno_generation}
\alias{geno_generation}
\title{Genotype Generation}
\usage{
geno_generation(snps, effect_snps_vec, n, maf)
}
\arguments{
\item{snps}{number of total snps}

\item{effect_snps_vec}{vector containing the number of effective additive,
dominant, recessive, and codominant snps}

\item{n}{number of individuals}

\item{maf}{minor allele frequency}
}
\value{
list consisting of:
\itemize{
\item dat_effect: dataframe containing effective snps in the columns and individuals in the rows
\item dat_effect_trans: similar with dat_effect
\item dat_non_effect: dataframe containing non-effective snps in the columns and individuals in the rows
}
}
\description{
Genotype Generation
}
