% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lasso_selection.R
\name{Lasso_selection}
\alias{Lasso_selection}
\title{Lasso Selection}
\usage{
Lasso_selection(
  prs_train = prs_train,
  dat_train = dat_train,
  theta_train = theta_train,
  family = "binomial",
  lambda = NULL
)
}
\arguments{
\item{prs_train}{dataframe containing the training polygenic risk scores}

\item{dat_train}{dataframe containing the training phenotype data}

\item{theta_train}{dataframe containing the training thetas}

\item{family}{specifies the type of model to fit, such as "binomial" for
binary outcomes and "gaussian" for continuous outcomes (default: "binomial")}

\item{lambda}{regularization parameter for Lasso regression (default: NULL)}
}
\value{
support: a vector indicating indices of selected SNPs from Lasso selection
}
\description{
This function performs Lasso regression to select important features
(SNPs or others) based on a given dataset and family of models.
}
