% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betas_thetas.R
\name{learning_betas}
\alias{learning_betas}
\title{Compute Beta Coefficients}
\usage{
learning_betas(
  pheno = pheno_train,
  geno = snp_train,
  prs = prs_train,
  cov = NULL,
  family = "binomial"
)
}
\arguments{
\item{pheno}{a dataframe containing the phenotype (outcome)}

\item{geno}{a datafrane containing the genotype data whose first column
stores the SNP rsID and second column stores the genotype encoding}

\item{prs}{a dataframe containing the polygenic risk scores}

\item{cov}{a dataframe containing the covariates}

\item{family}{a string naming the regression model used for glm. Use "binomial"
for binary outcomes and "gaussian" for continuous outcomes.}
}
\value{
a dataframe whose columns include the SNP rsID number, beta1, beta2,
and p-value.
}
\description{
This function computes beta coefficients and p-values of a generalized
linear model for each SNPs. The model fits the genotype, prs, and covariates
to the phenotype outcome. A binomial distribution should be used for binary
outcomes while a gaussian model should be used for continuous outcomes.
}
