% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{model_evaluation}
\alias{model_evaluation}
\title{Evaluation Function}
\usage{
model_evaluation(phenotype, ..., binary = TRUE)
}
\arguments{
\item{phenotype}{matrix containing the phenotype information}

\item{...}{dataframes that contain one set of polygenic risk scores each.
Please provide at least one set of polygenic risk scores. You can provide the
name of each set of polygenic risk score by providing a name for the parameter
in this format: \code{model_evaluation(phenotype=phenotype, 'Pre-trained PRS'=df1, 'ABPRS'=df2)}.}

\item{binary}{boolean that determines whether the data is binary or not}

\item{filename}{the name of the exported file}
}
\value{
An html file with the filename will appear in the current directory,
which contains different plots and data depending on whether the outcome is binary
or continuous.

For Binary Outcomes:
\itemize{
\item a downloadable performance score table with AUC scores for each set of PRSs
\item an interactable performance score comparison bar plot
\item an interactable polygenic risk score distribution density plot
\item an interactable percentage of cases vs polygenic risk score graph
\item an interactable odds ratio plot
\item the legend, ggplot code, and downloadable dataframe that produced the above plots
}

For Continuous Outcomes:
\itemize{
\item a downloadable performance score table with the Mean-Squared Error and \eqn{R^2}
for each set of PRSs
\item an interactable performance score comparison bar plot
\item an interactable mean phenotype vs polygenic risk score percentile graph
\item an interactable mean difference plot
\item the legend, ggplot code, downloadable dataframe that produced the above plots
}
}
\description{
This function evaluates the performance of at least one PRS model on the same
outcome by utilizing various methods, depending on whether the outcome is binary
or continuous.
}
