% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{Plot_Score}
\alias{Plot_Score}
\title{Barplot of AUC or Loss Score Comparison between ABPRS and PRS}
\usage{
Plot_Score(
  support_ABPRS = support_ABPRS,
  prs_train = prs_train,
  prs_test = prs_test,
  pheno_train = pheno_train,
  pheno_test = pheno_test,
  theta_train = theta_train,
  theta_test = theta_test,
  binary = TRUE
)
}
\arguments{
\item{support_ABPRS}{vector of indices specifying which the selected SNPS from ABPRS}

\item{prs_train}{dataframe containing the training polygenic risk scores}

\item{prs_test}{dataframe containing the testing polygenic risk scores}

\item{pheno_train}{dataframe containing the training phenotype data}

\item{pheno_test}{dataframe containing the testing phenotype data}

\item{theta_train}{dataframe containing the training thetas}

\item{theta_test}{dataframe containing the testing thetas}

\item{binary}{boolean specifying if the data is binary (TRUE) or
continuous (FALSE) (default: TRUE)}
}
\value{
Bar plot that compares the Area Under the Curve(binary data) or
Mean Squared Error (continuous data) score of the pre-trained PRS with that
of the ABPRS. See \link{AUC_test} for the AUC calculation and
\link{Loss_test} for the Mean Squared Error calculation methods.
}
\description{
Barplot of AUC or Loss Score Comparison between ABPRS and PRS
}
