% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{simulation_generation}
\alias{simulation_generation}
\title{Generate simulation data list necessary for data analysis}
\usage{
simulation_generation(snps, effect_snps_vec, n, maf, effect_size, beta0)
}
\arguments{
\item{snps}{number of total snps}

\item{n}{total number of samples}

\item{effect_size}{effect size of the effective snps}

\item{beta0}{beta0 value}

\item{effect_snp_vec}{vector of length 4 containing the number of effective snps
with additive, dominant, recessive, and heterozygous encoding respectively}
}
\value{
A list consisting of:
\itemize{
\item snps_order: snps that are ordered
\item theta_train, theta_val, theta_test
\item pheno_train, pheno_val, pheno_test
\item prs_train, prs_val, prs_test
\item geno_train, geno_val, geno_test
}
}
\description{
Generate simulation data list necessary for data analysis
}
