% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{split_data}
\alias{split_data}
\title{Split data}
\usage{
split_data(data, probability)
}
\arguments{
\item{data}{dataframe of original dataset}

\item{probability}{vector with the probability of each partition you wish to produce}
}
\value{
list of splited data
}
\description{
Split data
}
