% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABPRS.R
\name{ABPRS}
\alias{ABPRS}
\title{ABPRS Function}
\usage{
ABPRS(
  target_prs,
  target_genotype,
  training_genotype,
  validation_genotype,
  training_prs,
  validation_prs,
  training_phenotype,
  validation_phenotype,
  family = "binary",
  biglasso = FALSE,
  lam.max = 0.002,
  lam.min = 6e-05,
  nlambda = 50,
  alpha = 0.1,
  tolerance = 0.025,
  threshold = 0.01,
  err = 1e-05,
  delta = NULL
)
}
\arguments{
\item{target_prs}{dataframe containing the pre-trained prs that we wish to
convert}

\item{target_genotype}{dataframe containing the genotype information of the
target sample}

\item{training_genotype}{dataframe containing the training genotype data}

\item{validation_genotype}{dataframe containing the validation genotype data}

\item{training_prs}{dataframe containing the training polygenic risk scores}

\item{validation_prs}{dataframe containing the validation polygenic risk scores}

\item{training_phenotype}{dataframe containing the training phenotype data}

\item{validation_phenotype}{dataframe containing the validation phenotype data}

\item{family}{string specifying whether the data has binary or continuous outcomes.
Use "binary" for binary outcomes and "continuous" for continuous outcomes.
(default: "binary").}

\item{biglasso}{logical flag indicating whether to use big data optimization.
If true, then the program uses \link[biglasso]{biglasso}. Or else, it uses
the standard \link[glmnet]{glmnet} (default: FALSE)}

\item{lam.max}{maximum value of the regularization parameter (lambda) to be
considered. This value is used to generate the lambda sequence in \link[glmnet]{glmnet}.}

\item{lam.min}{minimum value of the regularization parameter (lambda) to be
considered. This value is used to generate the lambda sequence in \link[glmnet]{glmnet}.}

\item{nlambda}{number of different lambda values to be evaluated between
lam.max and lam.min. This value is used to generate the lambda sequence in \link[glmnet]{glmnet}.}

\item{alpha}{desired FDR control level.}

\item{tolerance}{tolerance level for noise in mirror statistics.}

\item{threshold}{threshold value for determining significance in mirror
statistics.}

\item{err}{error threshold for model performance improvement.}

\item{delta}{additional tuning parameter for the FDR control.}
}
\value{
Adaptively-boosted polygenic risk scores of the target sample.
}
\description{
This function adaptively-boosts pre-trained polygenic risk scores of a target
sample and returns the newly adaptively boosted polygenic risk scores of that
sample.
}
