% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{data_simulation}
\alias{data_simulation}
\title{Generate Simulation Data}
\usage{
data_simulation(
  snps_n,
  effect_snps_vec,
  sample_n,
  maf,
  effect_size,
  beta0,
  binary = TRUE
)
}
\arguments{
\item{snps_n}{total number of snps}

\item{sample_n}{total number of samples}

\item{effect_size}{effect size of the effective snps}

\item{beta0}{beta0 value, used to ensure a certain percentage of case phenotype}

\item{binary}{a boolean flag indicating whether the data has binary outcomes (TRUE)
or continuous outcomes (FALSE). (Default: TRUE)}

\item{effect_snp_vec}{vector of length 4 containing the number of effective snps
with additive, dominant, recessive, and codominant encoding respectively}
}
\value{
A dataframe containing the simulated phenotype in the first column and the
simulated genotype encoding in the rest of the columns.
}
\description{
This function generates simulated genotype and phenotype information for
\emph{snps_n} amount of snps and \emph{sample_n} amount of individuals.
}
