% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thetas.R
\name{learning_theta_snps}
\alias{learning_theta_snps}
\title{Compute \eqn{\theta_{SNPs}} for ABPRS}
\usage{
learning_theta_snps(
  phenotype,
  genotype,
  pre_trained_prs,
  covariate = NULL,
  family
)
}
\arguments{
\item{phenotype}{a dataframe containing the phenotype}

\item{genotype}{a dataframe containing the genotype matrix with additive encoding
whose columns represent SNPs and rows represent individuals.}

\item{pre_trained_prs}{a dataframe containing the pre-trained polygenic risk scores}

\item{covariate}{a dataframe containing covariates (default: NULL)}

\item{family}{a string specifying the type of model to fit a \link[stats]{glm}.
Use "binomial" for binary outcomes and "gaussian" for continuous outcomes.}
}
\value{
A dataframe whose columns include the SNP rsID, theta1 (theta encoding
for heterozygous alleles Aa), theta2 (theta encoding for homozygous alternative
alleles aa), and the p-value associated with \eqn{\theta_{SNPs}} and the outcome.
}
\description{
This function computes \eqn{\theta_{SNPs}} and their corresponding p-values for each SNP.
Specifically, the function fits a regression model to the phenotype outcome using the
genotype, PRS, and covariates. The function uses a binomial model for binary
outcomes and a Gaussian model for continuous outcomes.
}
